Add unit test to test_iss.py
This unit test is executed when we run 'make test' and handles the scenario when the user does not specify the '--p-sample-names' option. In the unit test, we call the 'generate_reads()' function without providing an element to the 'sample_names' list. In the previous version of the code, this unit test would fail, indicating the error message: 'Plugin error from assembly: 'NoneType' object is not iterable'.