BUG: correctly handle empty sample_names (#50)
* Added handling for empty or None sample_names

* iss.py - Correction

* Update iss.py

Add warning message for missing '--p-sample-names' option in verbose mode

* Add unit test to test_iss.py

This unit test is executed when we run 'make test' and handles the scenario when the user does not specify the '--p-sample-names' option. In the unit test, we call the 'generate_reads()' function without providing an element to the 'sample_names' list. In the previous version of the code, this unit test would fail, indicating the error message: 'Plugin error from assembly: 'NoneType' object is not iterable'.

* coding style corrections in iss.py

* coding style corrections in test_iss.py

* correction test_iss.py

* Code formatting correction for iss.py

* Adding quotes in the warning

* Update iss.py

* Update iss.py

* change of the position of the iss test

* redo test / isolating added functionality

* Remove Unnecessary Comments

* Apply suggestions from code review

---------

Co-authored-by: Michal Ziemski <mziemski@ethz.ch>